+++ pkg/storage/cached/private_test.go
@@ -1,4 +1,4 @@
-// Copyright 2022 The jackal Authors
+// Copyright 2021 The jackal Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -22,7 +22,7 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestCachedVCardRep_UpsertVCard(t *testing.T) {
+func TestCachedPrivateRep_UpsertPrivate(t *testing.T) {
 	// given
 	var cacheNS, cacheKey string
 
@@ -34,57 +34,55 @@ func TestCachedVCardRep_UpsertVCard(t *testing.T) {
 	}
 
 	repMock := &repositoryMock{}
-	repMock.UpsertVCardFunc = func(ctx context.Context, vcard stravaganza.Element, username string) error {
+	repMock.UpsertPrivateFunc = func(ctx context.Context, private stravaganza.Element, namespace string, username string) error {
 		return nil
 	}
 
 	// when
-	rep := cachedVCardRep{
+	rep := cachedPrivateRep{
 		c:   cacheMock,
 		rep: repMock,
 	}
-	vCard := stravaganza.NewBuilder("vCard").Build()
+	prv := stravaganza.NewBuilder("prv").Build()
 
-	err := rep.UpsertVCard(context.Background(), vCard, "u1")
+	err := rep.UpsertPrivate(context.Background(), prv, "n0", "u1")
 
 	// then
 	require.NoError(t, err)
-	require.Equal(t, vCardNS("u1"), cacheNS)
-	require.Equal(t, vCardKey, cacheKey)
-	require.Len(t, repMock.UpsertVCardCalls(), 1)
+	require.Equal(t, privateNS("u1"), cacheNS)
+	require.Equal(t, "n0", cacheKey)
+	require.Len(t, repMock.UpsertPrivateCalls(), 1)
 }
 
-func TestCachedVCardRep_DeleteVCard(t *testing.T) {
+func TestCachedPrivateRep_DeletePrivates(t *testing.T) {
 	// given
-	var cacheNS, cacheKey string
+	var cacheNS string
 
 	cacheMock := &cacheMock{}
-	cacheMock.DelFunc = func(ctx context.Context, ns string, keys ...string) error {
+	cacheMock.DelNSFunc = func(ctx context.Context, ns string) error {
 		cacheNS = ns
-		cacheKey = keys[0]
 		return nil
 	}
 
 	repMock := &repositoryMock{}
-	repMock.DeleteVCardFunc = func(ctx context.Context, username string) error {
+	repMock.DeletePrivatesFunc = func(ctx context.Context, username string) error {
 		return nil
 	}
 
 	// when
-	rep := cachedVCardRep{
+	rep := cachedPrivateRep{
 		c:   cacheMock,
 		rep: repMock,
 	}
-	err := rep.DeleteVCard(context.Background(), "v1")
+	err := rep.DeletePrivates(context.Background(), "u1")
 
 	// then
 	require.NoError(t, err)
-	require.Equal(t, vCardNS("v1"), cacheNS)
-	require.Equal(t, vCardKey, cacheKey)
-	require.Len(t, repMock.DeleteVCardCalls(), 1)
+	require.Equal(t, privateNS("u1"), cacheNS)
+	require.Len(t, repMock.DeletePrivatesCalls(), 1)
 }
 
-func TestCachedVCardRep_FetchVCard(t *testing.T) {
+func TestCachedPrivateRep_FetchPrivate(t *testing.T) {
 	// given
 	cacheMock := &cacheMock{}
 	cacheMock.GetFunc = func(ctx context.Context, ns, k string) ([]byte, error) {
@@ -95,25 +93,25 @@ func TestCachedVCardRep_FetchVCard(t *testing.T) {
 	}
 
 	repMock := &repositoryMock{}
-	repMock.FetchVCardFunc = func(ctx context.Context, username string) (stravaganza.Element, error) {
-		vCard := stravaganza.NewBuilder("vCard").Build()
-		return vCard, nil
+	repMock.FetchPrivateFunc = func(ctx context.Context, namespace string, username string) (stravaganza.Element, error) {
+		prv := stravaganza.NewBuilder("prv0").Build()
+		return prv, nil
 	}
 
 	// when
-	rep := cachedVCardRep{
+	rep := cachedPrivateRep{
 		c:   cacheMock,
 		rep: repMock,
 	}
-	vCard, err := rep.FetchVCard(context.Background(), "u1")
+	prv, err := rep.FetchPrivate(context.Background(), "n0", "u1")
 
 	// then
-	require.NotNil(t, vCard)
+	require.NotNil(t, prv)
 	require.NoError(t, err)
 
-	require.Equal(t, "vCard", vCard.Name())
+	require.Equal(t, "prv0", prv.Name())
 
 	require.Len(t, cacheMock.GetCalls(), 1)
 	require.Len(t, cacheMock.PutCalls(), 1)
-	require.Len(t, repMock.FetchVCardCalls(), 1)
+	require.Len(t, repMock.FetchPrivateCalls(), 1)
 }