+++ pkg/storage/cached/redis/cache.go
@@ -65,8 +65,9 @@ func New(cfg Config) *Cache {
 func (c *Cache) Type() string { return Type }
 
 // Get satisfies Cache interface.
-func (c *Cache) Get(ctx context.Context, k string) ([]byte, error) {
-	val, err := c.pickClient(k).Get(ctx, k).Result()
+func (c *Cache) Get(ctx context.Context, ns, key string) ([]byte, error) {
+	cl := c.pickClient(ns)
+	val, err := cl.HGet(ctx, ns, key).Result()
 	if err != nil {
 		if errors.Is(err, redis.Nil) {
 			return nil, nil
@@ -77,22 +78,34 @@ func (c *Cache) Get(ctx context.Context, k string) ([]byte, error) {
 }
 
 // Put satisfies Cache interface.
-func (c *Cache) Put(ctx context.Context, k string, val []byte) error {
-	return c.pickClient(k).Set(ctx, k, val, c.ttl).Err()
+func (c *Cache) Put(ctx context.Context, ns, key string, val []byte) error {
+	cl := c.pickClient(ns)
+	if err := cl.HSet(ctx, ns, key, val).Err(); err != nil {
+		return err
+	}
+	return cl.Expire(ctx, ns, c.ttl).Err()
 }
 
 // Del satisfies Cache interface.
-func (c *Cache) Del(ctx context.Context, k string) error {
-	return c.pickClient(k).Del(ctx, k).Err()
+func (c *Cache) Del(ctx context.Context, ns string, keys ...string) error {
+	cl := c.pickClient(ns)
+	return cl.HDel(ctx, ns, keys...).Err()
+}
+
+// DelNS removes all keys contained under a given namespace from the cache store.
+func (c *Cache) DelNS(ctx context.Context, ns string) error {
+	cl := c.pickClient(ns)
+	return cl.Del(ctx, ns).Err()
 }
 
 // HasKey satisfies Cache interface.
-func (c *Cache) HasKey(ctx context.Context, k string) (bool, error) {
-	v, err := c.pickClient(k).Exists(ctx, k).Result()
-	if err != nil {
+func (c *Cache) HasKey(ctx context.Context, ns, key string) (bool, error) {
+	cl := c.pickClient(ns)
+	res := cl.HExists(ctx, ns, key)
+	if err := res.Err(); err != nil {
 		return false, err
 	}
-	return v == 1, nil
+	return res.Val(), nil
 }
 
 // Start satisfies Cache interface.
@@ -115,11 +128,11 @@ func (c *Cache) Stop(_ context.Context) error {
 	return nil
 }
 
-func (c *Cache) pickClient(k string) *redis.Client {
+func (c *Cache) pickClient(ns string) *redis.Client {
 	if len(c.clients) == 1 {
 		return c.clients[0]
 	}
-	cs := xxhash.Sum64String(k)
+	cs := xxhash.Sum64String(ns)
 	idx := jumpHash(cs, len(c.clients))
 	return c.clients[idx]
 }