+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * [ENHANCEMENT] Cached VCard repository. #203
 * [ENHANCEMENT] Cached Last repository. #204
 * [ENHANCEMENT] Cached Capabilities repository. #205
+* [ENHANCEMENT] Cached Private repository. #206
 * [CHANGE] Introduced measured repository transaction type. #200
 * [CHANGE] Use PgSQL locker. #201
 * [BUGFIX] Fix S2S db key check when nop KV is used. #199